refactoring. models are in a separate file now
